+++ Exam/Exam.Domain/Model/ExamClass.cs
@@ -38,7 +38,7 @@ namespace Exam.Domain.Model
         /// <summary>
         /// --考试分数
         /// </summary>
-        public virtual int Score { get; set; }
+        public virtual decimal Score { get; set; }
 
         /// <summary>
         /// --排序
